Refactor start: moved old files in a single directory
