--- inhouse_bot/__init__.py
@@ -1,8 +1,0 @@
-from inhouse_bot.sqlite.sqlite_utils import get_session, initialize_sql
-from inhouse_bot.sqlite.game import Game
-from inhouse_bot.sqlite.game_participant import GameParticipant
-from inhouse_bot.sqlite.player import Player
-from inhouse_bot.sqlite.player_rating import PlayerRating
-from inhouse_bot.sqlite.queue_player import QueuePlayer
-
-initialize_sql()
