+++ inhouse_bot_old/__init__.py
@@ -1,0 +1,8 @@
+from inhouse_bot_old.sqlite.sqlite_utils import get_session, initialize_sql
+from inhouse_bot_old.sqlite.game import Game
+from inhouse_bot_old.sqlite.game_participant import GameParticipant
+from inhouse_bot_old.sqlite.player import Player
+from inhouse_bot_old.sqlite.player_rating import PlayerRating
+from inhouse_bot_old.sqlite.queue_player import QueuePlayer
+
+initialize_sql()
